refactor: 레이싱 게임의 오디오 제어 함수들을 utils로 빼서 모듈화 (CC-162)
